move PaneToggleButton to shared/Button/PaneToggle
