+++ src/components/SourceTabs.js
@@ -13,7 +13,7 @@ const { isEnabled } = require("devtools-config");
 const classnames = require("classnames");
 const actions = require("../actions");
 const CloseButton = require("./shared/CloseButton");
-const PaneToggleButton = require("./PaneToggleButton");
+const PaneToggleButton = require("./shared/Button/PaneToggle");
 const Svg = require("./utils/Svg");
 const Dropdown = React.createFactory(require("./Dropdown"));
 const { showMenu, buildMenu } = require("../utils/menu");
