+++ src/components/shared/Button/PaneToggle.js
@@ -1,8 +1,8 @@
 const React = require("react");
 const classnames = require("classnames");
-const Svg = require("./utils/Svg");
+const Svg = require("../../utils/Svg");
 
-require("./PaneToggleButton.css");
+require("./PaneToggle.css");
 
 function PaneToggleButton({
   position, collapsed, horizontal, handleClick, tooltip }) {
